Passed lab1

+++ lab1/ex4.cpp
@@ -17,21 +17,21 @@
     std::cout << std::endl;
 
 int main(){
-    int16_t xs=0xFADE;
+    short xs=0xFADE;
     int xl = 0xADE1A1DA;
-    long xq = 0xC1A551F1AB1E;
+    long long xq = 0xC1A551F1AB1E;
     float x1 = -8.0f/5.0f;
     double x2 = -8.0/5.0;
-    int16_t Ms[N]{xs,xs,xs,xs,xs};
+    short Ms[N]{xs,xs,xs,xs,xs};
     int Ml[N]{xl,xl,xl,xl,xl};
-    long Mq[N]{xq,xq,xq,xq,xq};
+    long long Mq[N]{xq,xq,xq,xq,xq};
     float Mfs[N]{x1,x1,x1,x1,x1};
     double Mfl[N]{x2,x2,x2,x2,x2};
 
 
-    PRINT_i(Ms,int16_t);
+    PRINT_i(Ms,short);
     PRINT_i(Ml,int);
-    PRINT_i(Mq,long);
+    PRINT_i(Mq,long long);
     PRINT_f(Mfs,float);
     PRINT_f(Mfl,double);
